% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_money.R
\name{has_outer_amount}
\alias{has_outer_amount}
\title{Detect presence of an outer amount}
\usage{
has_outer_amount(x)
}
\arguments{
\item{x}{A character vector containing monetary text.}
}
\value{
A logical vector indicating whether each element contains
a numeric value outside parentheses.
}
\description{
Determines whether a string contains a numeric value outside of any
trailing parentheses. This is used to decide whether a total amount
is explicitly provided.
}
\examples{
has_outer_amount("$1.75M CAD ($750k closed)")
has_outer_amount("($750k closed, $1M committed)")

}
