% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session_utils.R
\name{last_non_na}
\alias{last_non_na}
\title{Get last non-missing value}
\usage{
last_non_na(x)
}
\arguments{
\item{x}{A vector.}
}
\value{
A single value or NA if all values are missing.
}
\description{
Returns the last non-NA value in a vector.
}
